Split common visitor
